rename: Enum 클래스 이름 변경 - Role -> Authority
