+++ src/main/java/com/givemecon/domain/member/Member.java
@@ -1,7 +1,7 @@
 package com.givemecon.domain.member;
 
 import com.givemecon.config.enums.OAuth2Provider;
-import com.givemecon.config.enums.Role;
+import com.givemecon.config.enums.Authority;
 import com.givemecon.domain.BaseTimeEntity;
 import com.givemecon.domain.likedvoucher.LikedVoucher;
 import com.givemecon.domain.purchasedvoucher.PurchasedVoucher;
@@ -35,7 +35,7 @@ public class Member extends BaseTimeEntity {
 
     @Enumerated(EnumType.STRING)
     @Column(nullable = false, length = 10)
-    private Role role;
+    private Authority authority;
 
     @Enumerated(EnumType.STRING)
     @Column(length = 10)
@@ -48,18 +48,18 @@ public class Member extends BaseTimeEntity {
     List<PurchasedVoucher> purchasedVoucherList = new ArrayList<>();
 
     @Builder
-    public Member(String email, String username, String password, Role role) {
+    public Member(String email, String username, String password, Authority authority) {
         this.email = email;
         this.username = username;
         this.password = password;
-        this.role = role;
+        this.authority = authority;
     }
 
     @Builder(builderClassName = "oauthBuilder", builderMethodName = "oauthBuilder")
-    public Member(String email, String username, Role role, OAuth2Provider provider) {
+    public Member(String email, String username, Authority authority, OAuth2Provider provider) {
         this.email = email;
         this.username = username;
-        this.role = role;
+        this.authority = authority;
         this.provider = provider;
     }
 
@@ -70,8 +70,8 @@ public class Member extends BaseTimeEntity {
         return this;
     }
 
-    public String getRoleKey() {
-        return role.getKey();
+    public String getRole() {
+        return authority.getRole();
     }
 
     public void addLikedVoucher(LikedVoucher likedVoucher) {
