+++ src/test/java/com/givemecon/controller/DtoValidationTest.java
@@ -24,7 +24,7 @@ import java.nio.charset.StandardCharsets;
 import java.util.List;
 
 import static com.givemecon.config.enums.JwtAuthHeader.*;
-import static com.givemecon.config.enums.Role.*;
+import static com.givemecon.config.enums.Authority.*;
 import static com.givemecon.controller.TokenHeaderUtils.*;
 import static com.givemecon.domain.member.MemberDto.*;
 import static com.givemecon.util.error.ErrorCode.*;
@@ -65,7 +65,7 @@ public class DtoValidationTest {
         member = memberRepository.save(Member.builder()
                 .username("tester")
                 .email("test@gmail.com")
-                .role(ADMIN)
+                .authority(ADMIN)
                 .build());
 
         tokenInfo = jwtTokenService.getTokenInfo(new TokenRequest(member));
